RWC-64 Updates:
Prepared code for legacy RCO.

Injected $ResursCheckout.release(); into v2 as a finalization at the side.

Fixed resursBankRcoDataContainer-blocks.

Legacy fieldchecker is using getCustomerFieldsByApiVersion to merge everything properly.

Legacy auto detection for legacy in generic isLegacyIframe.

Running in RCO mode, we will only have one method available and therefore we change the current id to that method.

getOrder is used to fetch the WC_Order globally as soon as it has been created and is initialized.

Added support to move the iframe around in different checkout sections.

Applied ajax_success to the ajaxifier, or the checkout won't react at the front page.

Added support in order processing for RCO so that we won't pass through the generic process_payment that is used internally by WooCommerce. When we run iframed, the order has been precreated already, but in the backend.

Updated processRco with order processing. Also added data about legacy in the session, so it can be reused by customers in the finalization end of the payment.

Added legacy script in enqueuer.

Changed some order status constants as they are changed and deprecated in ecomphp.
   -            RESURS_PAYMENT_STATUS_RETURNCODES::PAYMENT_PROCESSING => 'processing',
   +            OrderStatus::PROCESSING => 'processing',
And so on.

Added in the js autoloader (for legacy):
      'resursbank_rco_v1' => 'resursbank_rco_v1.js',
      'resursbank_rco_v2' => 'resursbank_rco_v2.js',
Also added dependencies to jquery.

Added function to pick up error notices generated during order creation in RCO.

Update PluginApi to support backend order creations (rco only). Using a merger during the _POST handler to merge RCO customer data in the "real" form field handling (overwriting the regular post fields). On the other side of the processing, we pick up the current order again with order_awaiting_payment.

Created a plugin class to use when handling own internal customizations. This will be set to create snippets in the docs too further on.
This one is used internally in the order processor methods and can also be used by other developers with similar needs (when it comes to skipping the internal process).
Both filters are required for everything to kick in properly.